Revamp installer package for Leopard.
Lots of changes to the build process for the installer package. Apple totally
overhauled the PackageMaker infrastructure again. I have now switched to a
new-style flat package that contains everything in one file. The package
description is now generated by the PackageMaker utility but the package itself
can still be build from the commandline. This makes the whole process a whole
lot easier.